pylib: extract upgrade helpers from test_sstable_compression_dictionaries_upgrade.py
The cherry-picked patch was originally introduced in 2025.4 to reuse the
helpers for testing upgrades for LWT fencing (368d70ee15).

Later, it was also used to test the default SSTable compression
algorithm (adf9c426c2).

Backport this patch so that we can later backport the PR for the default
SSTable compression algorithm.

(cherry picked from commit 49b036cf2bcd8582eaf55517fdd2c1f3c4531555)
